backend/requirements.txt
fastapi
uvicorn
python-multipart
librosa
soundfile
torch
tensorflow
numpy
pillow
transformers
datasets
openai
python-dotenv
scipy
opencv-python-headless
PyJWT[crypto]
cryptography
requests
edge-tts==7.2.7
huggingface_hub


# Testing
pytest
httpx
pytest-asyncio
pytest-cov
starlette==0.50.0
pydantic==2.11.10
fastapi==0.121.3
